Initial commit of receive_chain module. Work in progress.
An intial attempt at the receive chain module is now complete. There are
some bugs to work out ( i.e. the output is not yet correct ). Due to the
wide range of changes made across files, all module-level tests will
need to be re-evaluated. This commit is more of an attempt to keep
progress manageable.