More progress on the rx chain module.
Determined that the Halfband filter is producing the wrong output. I've
simply bypassed that for now, so the output comes directly from the CIC
filter instead ( this is optional in the USRP ). The test case fir
filter works so there must be a bug in the half band implementation.

All makefiles have now been removed.

Wrote a test case for the sinusoid generator class.

TODO:

1. Implement 2nd half of rx chain to handle complex outputs.
2. Fix the half-band filter bug.
3. Work on a new implementation of the existing USRP hbf.
4. Refactor the abstract FirFilter class into something that makes more
sense.
5. Convert octave scripts to python equivalents.
6. Create a more complex signal generator to simulate meteor events.
7. Create more documentation for basic use cases.