Fixed cordic theta map bug that was re-introduced during cleanup.
The cordic theta mapper is used to remap signal vectors that fall into
the 2nd or 3rd quadrants. This is required since arctan is only valid in
the interval (-pi/2,pi/2). We must also remap the performance variable,
which happens by clipping the MSB from the original. I was clipping the
LSB, which was not flipping the variable when required. Verified through
testing in the cordic folder.