Added XML configuration capability and a generic code generator class.
Since systemc uses templates to define data types, all defined widths
must be passed as compile-time constants, thus the reason for creating a
code generator class. Also replaced boost jam with scons as the primary
build system and created dependencies ( so far this experiment is
limited to the accumulator code ) that enable the build system to
perform the following tasks:

1. Parse the xml configuration file.
2. Produce a header file containing constants read from the xml file.
3. Compile the module specific code with the newly defined header.

TODO:

Replace all modules with this new system.