Fixed bit-pruning gain issue and updated octave scripts.
Gain normalization for bit-pruned CIC designs requires gain calculations using post-pruning bit width of the first stage, as opposed to the more obvious calculation of N*log2(Rmax/Rselected). The final gain could be slightly different across different values of decimation due to non-integer Rmax calculations resulting from post-pruned bit width. This type of calculation would require pre-computation in the form of a look-up table along with a bit-shifting stage if used in hardware. This sort of thing is NOT discussed in the famous hogenauer paper, but is extremely important if you want to build a variable-decimation CIC filter that includes pruning.